Separate directories for pages and components
