+++ src/App.jsx
@@ -1,18 +1,11 @@
 import React from 'react';
 import './css/App.css';
 
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Link
-} from "react-router-dom";
+import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
 
-import LandingPage from './components/LandingPage';
-import LoginPage from './components/LoginPage';
-import ContactBook from './components/ContactBook';
-
-import Login from './Login';
+import LandingPage from './pages/LandingPage';
+import LoginPage from './pages/LoginPage';
+import ContactBook from './pages/ContactBook';
 
 function App() {
   return (
@@ -28,29 +21,29 @@ function App() {
       <body>
         <h1 className="jumbotron-heading">App main page</h1>
         <Router>
-      <div>
-        <nav>
-          <ul>
-            <li>
-              <Link to="/login">Log in here</Link>
-            </li>
-            <li>
-              <Link to="/landing">Landing</Link>
-            </li>
-            <li>
-              <Link to="/contact-book">Contact Book</Link>
-            </li>
-          </ul>
-        </nav>
-        {/* A <Switch> looks through its children <Route>s and
+          <div>
+            <nav>
+              <ul>
+                <li>
+                  <Link to="/login">Log in here</Link>
+                </li>
+                <li>
+                  <Link to="/landing">Landing</Link>
+                </li>
+                <li>
+                  <Link to="/contact-book">Contact Book</Link>
+                </li>
+              </ul>
+            </nav>
+            {/* A <Switch> looks through its children <Route>s and
             renders the first one that matches the current URL. */}
-        <Switch>
-          <Route path="/login" exact component={LoginPage}/>
-          <Route path="/landing" exact component={LandingPage}/>
-          <Route path="/contact-book" exact component={ContactBook}/>
-        </Switch>
-      </div>
-    </Router>
+            <Switch>
+              <Route path="/login" exact component={LoginPage} />
+              <Route path="/landing" exact component={LandingPage} />
+              <Route path="/contact-book" exact component={ContactBook} />
+            </Switch>
+          </div>
+        </Router>
       </body>
     </div>
   );
