--- src/components/LoginPage.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function LoginPage() {
-    return (
-        <h1>Login Page</h1>
-  );
-}
