+++ src/pages/LoginPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import Login from '../components/Login';
+
+function LoginPage() {
+  return <Login />;
+}
+
+export default LoginPage;
